Add keymap descriptions
This is very useful for integration with [which-key](https://github.com/folke/which-key.nvim).
By default, which-key would describe a mapping with the function call -
e.g. `require("trouble").action("open_vsplit")`, which is not too bad to read,
but often truncated to e.g. `require("trouble").action("op`.
With this patch, the which-key description reads `Trouble open_vsplit`
instead.