Make the base LSP label more like others
The base entry had an label with the LSP client name in
parentheses, which pushes the range to the right, making
it more likely that the range will leave the edge of the viewport.

Since the range is more useful information than the LSP client,
change the base label format to be more like the others.

This also improves the appearance of the tree.

e.g.

```
     └╴1 ACCOUNT-ENTRY-VIEW VIEW OF DLP-ACCOUNT-ENTRY (natls) [105, 3]

     becomes                                      highlighted  v---v

     └╴1 ACCOUNT-ENTRY-VIEW VIEW OF DLP-ACCOUNT-ENTRY [105, 3] natls
```